refactor: extract pipeline.py and unify static serving (quality plan 7-8)
§8: Extract clip processing functions (process_clips, process_reel,
compute_reel_id, regenerate_from_manifest, is_excel_worksheet) from
clipgen.py into pipeline.py so server.py no longer depends on the CLI
entry point. Backward-compatible re-exports preserved in clipgen.py.

§7: Add utils.register_static_routes() factory and replace 4
independent serve_index/serve_static/serve_icons/serve_media
implementations across server.py, insights_server.py,
screenspace_server.py, and transcripts_server.py.
