+++ utils.py
@@ -1278,3 +1278,56 @@ def discover_participant_videos(study_name: str = "") -> list[dict[str, Any]]:
                     }
                 )
     return participants
+
+
+# ---- Flask blueprint helpers ----
+
+
+def register_static_routes(
+    bp: Any,
+    index_html: str,
+    *,
+    media_dir_getter: Any = None,
+    media_error: str = "Media directory not configured",
+    icons: bool = False,
+) -> None:
+    """Register standard static-file serving routes on a Flask Blueprint.
+
+    Always registers ``/`` (index) and ``/<path:filename>`` (static assets).
+    Optionally registers ``/icons/<path:filename>`` and ``/media/<path:filename>``.
+
+    Args:
+        bp: Flask Blueprint to register routes on.
+        index_html: Filename of the HTML page served at ``/``.
+        media_dir_getter: Callable returning the current media directory path.
+            When provided, a ``/media/<path:filename>`` route is registered.
+        media_error: Error message returned (500) when the media dir is falsy.
+        icons: When True, registers ``/icons/<path:filename>`` from ``assets/icons/``.
+    """
+    from flask import Response, jsonify, send_from_directory
+
+    assets_dir = get_bundled_assets_root() / "assets" / "web"
+
+    @bp.route("/")
+    def serve_index() -> Response:
+        return send_from_directory(assets_dir, index_html)
+
+    @bp.route("/<path:filename>")
+    def serve_static(filename: str) -> Response:
+        return send_from_directory(assets_dir, filename)
+
+    if icons:
+        icons_dir = get_bundled_assets_root() / "assets" / "icons"
+
+        @bp.route("/icons/<path:filename>")
+        def serve_icons(filename: str) -> Response:
+            return send_from_directory(icons_dir, filename)
+
+    if media_dir_getter is not None:
+
+        @bp.route("/media/<path:filename>")
+        def serve_media(filename: str) -> Response | tuple[Response, int]:
+            d = media_dir_getter()
+            if not d:
+                return jsonify({"ok": False, "error": media_error}), 500
+            return send_from_directory(d, filename)
